all but payment done

+++ lib/features/admin/services/admin_services.dart
@@ -20,7 +20,7 @@ class AdminServices {
     required String brand,
     required String description,
     required double price,
-    required double quantity,
+    required int quantity,
     required String? category,
     required List<File> images,
   }) async {
